backend/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
Flask-Limiter==3.0.0
python-dotenv==1.0.0
markdown2==2.4.8
Werkzeug==2.3.6
gunicorn==21.2.0
